Invitation/new user oauth signup operational
A user (currently just authors) be invited to corere via a link. When they use the link they are able to associate an oauth account of theirs to a "blank" account with the right permissions. They are then asked to fill out more info about themselves before interacting with corere